[IMP]project_task_milestone: fields restated

+++ project_task_milestone/models/project_task_milestone.py
@@ -1,7 +1,7 @@
 # Copyright 2018 Gontzal Gomez - AvanzOSC
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
-from odoo import api, fields, models
+from odoo import fields, models
 
 
 class ProjectTask(models.Model):
@@ -10,7 +10,4 @@ class ProjectTask(models.Model):
     milestone = fields.Boolean(string='Milestone')
     phase_id = fields.Many2one(
         string='Phase', comodel_name='project.task.phase')
-    sequence = fields.Float(string='Sequence') # compute='_commpute_sequence')
-
-    # @api.multi
-    # def _compute_sequence(self):
+    sequence = fields.Integer(string='Sequence')
